refactor(crossover trainer): 优化分布式GA训练的统计聚合逻辑，统一浮点精度
1.  将多处np.float64替换为np.float32以降低内存占用
2.  重构crossover_nft.py中的分布式reduce统计代码，使用有序key列表确保跨进程打包一致性
3.  简化统计构建流程，合并原_build_ga_stats逻辑到主函数中

+++ src/flow_factory/trainers/crossover/genetic_algorithm.py
@@ -547,7 +547,7 @@ class GeneticAlgorithm:
 
         # Pareto mask
         stack = np.stack(
-            [combined_rewards[k].astype(np.float64) for k in reward_keys], axis=1
+            [combined_rewards[k].astype(np.float32) for k in reward_keys], axis=1
         )
         pareto = compute_pareto_mask(stack)
 
@@ -649,10 +649,10 @@ class GeneticAlgorithm:
         if n == 0:
             return np.array([])
 
-        agg = np.zeros(n, dtype=np.float64)
+        agg = np.zeros(n, dtype=np.float32)
         for key in reward_keys:
             w = next(iter(self._reward_weights.get(key, {"default": 1.0}).values()))
-            vals = rewards_dict[key].astype(np.float64)
+            vals = rewards_dict[key].astype(np.float32)
             mean = vals.mean()
             std = vals.std()
             if std > 1e-8 and n > 1:
